python/nms/requirements.txt
python-dateutil==2.8.1
requests==2.25.1
sqlalchemy==1.3.19
mysql-connector-python==8.0.25
pymysql==0.10.1
celery==4.4.7
redis==3.5.3
cryptography==3.4.7
